+++ internal/cli/cli.go
@@ -7,6 +7,7 @@ import (
 
 	cli "github.com/urfave/cli/v2"
 
+	"github.com/mvisonneau/vac/internal/cli/flags"
 	"github.com/mvisonneau/vac/internal/cmd"
 )
 
@@ -28,37 +29,11 @@ func NewApp(version string, start time.Time) (app *cli.App) {
 	app.EnableBashCompletion = true
 
 	app.Flags = cli.FlagsByName{
-		&cli.StringFlag{
-			Name:    "engine",
-			Aliases: []string{"e"},
-			EnvVars: []string{"VAC_ENGINE"},
-			Usage:   "engine `path`",
-		},
-		&cli.StringFlag{
-			Name:    "role",
-			Aliases: []string{"r"},
-			EnvVars: []string{"VAC_ROLE"},
-			Usage:   "role `name`",
-		},
-		&cli.StringFlag{
-			Name:    "state",
-			Aliases: []string{"s"},
-			EnvVars: []string{"VAC_STATE_PATH"},
-			Usage:   "state `path`",
-			Value:   "~/.vac_state",
-		},
-		&cli.StringFlag{
-			Name:    "log-level",
-			EnvVars: []string{"VAC_LOG_LEVEL"},
-			Usage:   "log `level` (debug,info,warn,fatal,panic)",
-			Value:   "info",
-		},
-		&cli.StringFlag{
-			Name:    "log-format",
-			EnvVars: []string{"VAC_LOG_FORMAT"},
-			Usage:   "log `format` (json,text)",
-			Value:   "text",
-		},
+		flags.Engine,
+		flags.LogFormat,
+		flags.LogLevel,
+		flags.Role,
+		flags.State,
 	}
 
 	app.Action = cmd.ExecWrapper(cmd.Switch)
@@ -69,25 +44,9 @@ func NewApp(version string, start time.Time) (app *cli.App) {
 			Usage:  "get the creds in credential_process format (json)",
 			Action: cmd.ExecWrapper(cmd.Get),
 			Flags: cli.FlagsByName{
-				&cli.DurationFlag{
-					Name:    "min-ttl",
-					EnvVars: []string{"VAC_MIN_TTL"},
-					Usage:   "min-ttl `duration`",
-					Value:   0,
-				},
-				&cli.DurationFlag{
-					Name:    "ttl",
-					Aliases: []string{"t"},
-					EnvVars: []string{"VAC_TTL"},
-					Usage:   "ttl `duration`",
-					Value:   0,
-				},
-				&cli.BoolFlag{
-					Name:    "force-generate",
-					Aliases: []string{"f"},
-					EnvVars: []string{"VAC_FORCE_GENERATE"},
-					Usage:   "bypass currently cached creds and generate new ones",
-				},
+				flags.ForceGenerate,
+				flags.MinTTL,
+				flags.TTL,
 			},
 		},
 		{