+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [0ver](https://0ver.org) (more or less).
 
 - Golang updated to `1.20`
 - Bumped all dependencies
-- 
+- internal/cli: enhanced flags implementation
+
 ## [v0.0.8] - 2021-11-15
 
 ### Added